Move LanguageCodesSetupMixin to separate file
